Remove Adult dataset, auto download/process from source.
