--- THIRD-PARTY
@@ -1,65 +1,0 @@
-** UCI Adult; version 1.0 -- https://archive.ics.uci.edu/ml/datasets/Adult
-Copyright UC Irvine
-
-Reproduced:
-
-If you publish material based on databases obtained from this repository, then,
-in your acknowledgements, please note the assistance you received by using this
-repository. This will help others to obtain the same data sets and replicate
-your experiments. We suggest the following pseudo-APA reference format for
-referring to this repository:
-
-Dua, D. and Graff, C. (2019). UCI Machine Learning Repository
-[http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School
-of Information and Computer Science.
-
-Here is a BiBTeX citation as well:
-
-@misc{Dua:2019 ,
-author = "Dua, Dheeru and Graff, Casey",
-year = "2017",
-title = "{UCI} Machine Learning Repository",
-url = "http://archive.ics.uci.edu/ml",
-institution = "University of California, Irvine, School of Information and
-Computer Sciences" }
-
-A few data sets have additional citation requests. These requests can be found
-on the bottom of each data set's web page.
-
-------
-
-** scikit-learn; version 0.21.2 -- https://pypi.org/project/scikit-learn/
-Copyright (c) 2007–2019 The scikit-learn developers.
-All rights reserved.
-
-New BSD License
-
-Copyright (c) 2007–2019 The scikit-learn developers.
-All rights reserved.
-
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are met:
-
-  a. Redistributions of source code must retain the above copyright notice,
-     this list of conditions and the following disclaimer.
-  b. Redistributions in binary form must reproduce the above copyright
-     notice, this list of conditions and the following disclaimer in the
-     documentation and/or other materials provided with the distribution.
-  c. Neither the name of the Scikit-learn Developers  nor the names of
-     its contributors may be used to endorse or promote products
-     derived from this software without specific prior written
-     permission.
-
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-ARE DISCLAIMED. IN NO EVENT SHALL THE REGENTS OR CONTRIBUTORS BE LIABLE FOR
-ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
-LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY
-OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH
-DAMAGE.
